requirements.txt
numpy
opencv-python
tqdm
matplotlib
scikit-learn
pillow